+++ packages/hydrogen/src/foundation/Cookie/tests/Cookie.test.ts
@@ -1,4 +1,3 @@
-import {Logger} from '../../../utilities/log';
 import {Cookie} from '../Cookie';
 import {log} from '../../../utilities/log';
