+++ packages/hydrogen/src/foundation/useSession/useSession.ts
@@ -1,0 +1,7 @@
+import {useServerRequest} from '../ServerRequestProvider';
+
+export const useSession = function () {
+  const request = useServerRequest();
+  const session = request.ctx.session?.get();
+  return session;
+};
