+++ packages/hydrogen/src/index.ts
@@ -44,7 +44,7 @@ export {
 export {fetchSync} from './foundation/fetchSync/server/fetchSync';
 export {useServerAnalytics} from './foundation/Analytics';
 
-export {useSession} from './foundation/session/session';
-export {CookieSessionStorage} from './foundation/session/CookieSessionStorage';
-export {MemorySessionStorage} from './foundation/session/MemorySessionStorage';
-export {Cookie} from './foundation/session/Cookie';
+export {useSession} from './foundation/useSession/useSession';
+export {CookieSessionStorage} from './foundation/CookieSessionStorage/CookieSessionStorage';
+export {MemorySessionStorage} from './foundation/MemorySessionStorage/MemorySessionStorage';
+export {Cookie} from './foundation/Cookie/Cookie';
