--- packages/hydrogen/src/node.ts
@@ -1,1 +1,0 @@
-export {FileSessionStorage} from './foundation/session/FileSessionStorage';
