+++ packages/hydrogen/src/utilities/apiRoutes.ts
@@ -207,6 +207,10 @@ export async function renderApiRoute(
       }
     }
 
+    if (!response) {
+      response = new Response(null);
+    }
+
     if (cookieToSet) {
       response.headers.set('Set-Cookie', cookieToSet);
     }
